src/airflow/requirements.txt
apache-airflow==2.10.4
apache-airflow-providers-postgres
apache-airflow-providers-docker
apache-airflow-providers-apache-kafka
pandas
elasticsearch
psycopg[binary]
pyspark==3.5.6
apache-airflow-providers-apache-spark==5.2.1
pymysql==1.1.0
sshtunnel==0.4.0
google-cloud-bigquery==3.21.0
pyarrow==15.0.2
dbt-core==1.7.9
dbt-bigquery==1.7.9